Moving commons to the root of the project
